% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adam-data}
\alias{adam-data}
\alias{ADSL}
\alias{adsl}
\alias{ADAE}
\alias{adae}
\alias{adlb}
\alias{ADLB}
\title{ADAM example data}
\description{
Subject Level (ADSL), Lab (ADLB) and Adverse Events (ADAE) datasets used in
NNR examples.
}
\examples{
data(adsl)
print(adsl)

data(adae)
print(adae)

data(adlb)
print(adlb)
}
\keyword{data}
