% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_add.R
\name{addAlignment}
\alias{addAlignment}
\title{Add alignment specifications to an .NNTable}
\usage{
addAlignment(.NNTable, ...)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{...}{The alignment specification. It can be specified as: some_col1 =
"c", "some with col2" = "l", some_col3 = "r", where some_col1, "some with
col2", and some_col3 corresponds to the column names in the original data.
When used together with \code{link{addTranLong}} the created column names
need to be used.}
}
\value{
An object of class \code{NNTable} with the formatting applied
}
\description{
Add alignment specifications to an .NNTable
}
\examples{
## *****************************************************************************
## # Example of transpose from wide to long                                 ----
##
## Calculate the statistics for an AE table
## *****************************************************************************

library(dplyr)
# library(NNR)

# Filter data according Safety analysis set
adsl_f <- adsl            \%>\%
  filter(SAFFL == "Y")    \%>\%
    rename(TRTA = TRT01P)

# Calculate total exposure
totals <- adsl_f        \%>\%
  group_by(TRTA)        \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate adverse event statistics
output <- adae \%>\% group_by(AEBODSYS, AEDECOD, TRTA) \%>\%
  dplyr::summarise(N = n_distinct(USUBJID),
                   P = N / nrow(adsl),
                   E = n())

# Generate the NNTable ------------------------------------------------------

# Initiate by defining the columns based on ADAE
.NNTable <- NNTable(output, "AEBODSYS", "AEDECOD", "TRTA", "N", "(\%)" = "(P)", "E")

# Add exposure calculated from ADSL
.NNTable <- .NNTable \%>\%
  addExposure(exposure = totals,
              format_alone = FALSE,
              format_data  = c("Number of subjects"     = "\%.0f",
                               "Total exposure (years)" = "\%.2f"))

# Test alignment
.NNTable <- .NNTable \%>\% addAlignment(AEDECOD = "c", AEBODSYS = "r")

# print the table
.NNTable
}
