% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCellSplit.R
\name{addCellSplit}
\alias{addCellSplit}
\title{Split cells into multiple cells in accordance with a splitter}
\usage{
addCellSplit(.NNTable, split = "\\\\|", align = c("top", "bottom", "centre"))
}
\arguments{
\item{.NNTable}{An \code{NNTable}}

\item{split}{\code{character} specifying the character used to split}

\item{align}{\code{character} specifying alignment of splits}
}
\value{
Return the NNTable with the the truncation specified
}
\description{
Split cells into multiple cells in accordance with a splitter
}
