% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_total.R
\name{addExposure}
\alias{addExposure}
\title{Add exposure information to the top of the table}
\usage{
addExposure(.NNTable, exposure, ..., format_alone = FALSE, format_data = NULL)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{exposure}{data.frame with the exposure information to add}

\item{...}{rename columns}

\item{format_alone}{\code{logical} Should the exposure be formatted alone or
together with the column it is assigned to}

\item{format_data}{A vector or data.frame specifying the format for each
variable}
}
\value{
An object of class \code{NNTable} with the exposure data added
}
\description{
Add exposure information such as total number of subjects and years of
exposure
}
\examples{
## *****************************************************************************
## # Calculate summary statistics                                           ----
##
## The exposure/summary statistics are calculated for ADSL and ADLB
## *****************************************************************************

# filter and rename adsl to get totals
library(dplyr)

adsl_f <- adsl       \%>\%
filter(FASFL == "Y") \%>\%
  rename(TRTP = TRT01P)

# calculate the totals by treatment arm
totals <- adsl_f         \%>\%
  mutate(TRTP = "Total") \%>\%
  bind_rows(adsl_f)      \%>\%
  filter(FASFL == "Y")   \%>\%
  group_by(TRTP)         \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate the summary statistcs
output <- adlb                                                  \%>\%
  mutate(TRTP = "Total")                                        \%>\%
  bind_rows(adlb)                                               \%>\%
  filter(SAFFL == "Y" & PARAMCD \%in\% c("ABSR148N", "ABTIT148")) \%>\%
  group_by(TRTP, PARAM, STUDYID)                                \%>\%
  summarize(N  = n_distinct(USUBJID),
            Mean = mean(AVAL, na.rm = TRUE),
            SD   = sd(  AVAL, na.rm = TRUE),
            Min  = min( AVAL, na.rm = TRUE),
            Max  = max( AVAL, na.rm = TRUE)
  )

# The NNTable is generated in small steps to view what happens

# Initiate the table
.NNTable <- NNTable(output, "PARAM", "TRTP", "N", "Mean (SD)", "Min ; Max")

# View the table
.NNTable


# Transpose the summary columns to long format
.NNTable <- .NNTable                                    \%>\%
  addTransLong(SUM = c("N", "Mean (SD)" , "Min ; Max"),
               var_name = "Called var")                 \%>\%
  addFormat(format_data = c(N = "\%.0f"), dec = 2)

# View the table
.NNTable

# Add exposure, transpose to wide and ad grouping variables
.NNTable <- addExposure(.NNTable, exposure = totals,
                        format_alone = TRUE,
                        format_data  = c("Number of subjects" = "\%.0f")) \%>\%
            addTransWide(TRTP = "SUM")                                   \%>\%
            addGroupedColumns("PARAM", "Called var")

# View the final result
.NNTable
}
