% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFormat.R
\name{addFormat}
\alias{addFormat}
\title{Add formatting info to an NNTable}
\usage{
addFormat(
  .NNTable,
  dec = 3,
  format_data = NULL,
  big.mark = "",
  small.mark = "",
  group_by = NULL,
  ...
)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{dec}{\code{integer} defining the number of decimals to use}

\item{format_data}{\code{data.frame} containing a set of formats. The column
names of the supplied data to \code{\link{NNTable}} will be matched to the
column names of \code{format_data}.}

\item{big.mark}{\code{character} for thousands Separator}

\item{small.mark}{\code{character} for Separating each fifth decimal}

\item{group_by}{\code{character} with golumns indicating groups by which the
format is calculated individually}

\item{...}{Additional arguments, not currently supported}
}
\value{
An object of class \code{NNTable} with the formatting applied
}
\description{
Add formatting info to an NNTable
}
\examples{
## *****************************************************************************
## # Calculate summary statistics                                           ----
##
## The exposure/summary statistics are calculated for ADSL and ADLB
## *****************************************************************************

# filter and rename adsl to get totals
# library(NNR)
library(dplyr)

adsl_f <- adsl       \%>\%
filter(FASFL == "Y") \%>\%
  rename(TRTP = TRT01P)

# calculate the totals by treatment arm
totals <- adsl_f         \%>\%
  mutate(TRTP = "Total") \%>\%
  bind_rows(adsl_f)      \%>\%
  filter(FASFL == "Y")   \%>\%
  group_by(TRTP)         \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate the summary statistcs
output <- adlb                                                  \%>\%
  mutate(TRTP = "Total")                                        \%>\%
  bind_rows(adlb)                                               \%>\%
  filter(SAFFL == "Y" & PARAMCD \%in\% c("ABSR148N", "ABTIT148")) \%>\%
  group_by(TRTP, PARAM, STUDYID)                                \%>\%
  summarize(N  = n_distinct(USUBJID),
            Mean = mean(AVAL, na.rm = TRUE),
            SD   = sd(  AVAL, na.rm = TRUE),
            Min  = min( AVAL, na.rm = TRUE),
            Max  = max( AVAL, na.rm = TRUE)
  )

# The NNTable is generated in small steps to view what happens

# Initiate the table
.NNTable <- NNTable(output, "PARAM", "TRTP", "N", "Mean (SD)", "Min ; Max")

# View the table
.NNTable


# Transpose the summary columns to long format
.NNTable <- .NNTable                                    \%>\%
  addTransLong(SUM = c("N", "Mean (SD)" , "Min ; Max"),
               var_name = "Called var")                 \%>\%
  addFormat(format_data = c(N = "\%.0f"), dec = 2)

# View the table
.NNTable

# Add exposure, transpose to wide and ad grouping variables
.NNTable <- addExposure(.NNTable, exposure = totals,
                        format_alone = TRUE,
                        format_data  = c("Number of subjects" = "\%.0f")) \%>\%
            addTransWide(TRTP = "SUM")                                   \%>\%
            addGroupedColumns("PARAM", "Called var")

# View the final result
.NNTable
}
