% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPrinter.R
\name{addPrinter}
\alias{addPrinter}
\title{Specify print format.}
\usage{
addPrinter(.NNTable, type = c("txt", "flextable", "docx"), ...)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{type}{\code{character} dictating the ouput format}

\item{...}{Additional arguments for the specific print method}
}
\value{
An object of class NNTable with the printer specified
}
\description{
Specify print format.
}
\examples{
## *****************************************************************************
## # Calculate summary statistics                                           ----
##
## The exposure/summary statistics are calculated for ADSL and ADLB
## *****************************************************************************

# filter and rename adsl to get totals
# library(NNR)
library(dplyr)

adsl_f <- adsl       \%>\%
filter(FASFL == "Y") \%>\%
  rename(TRTP = TRT01P)

# calculate the totals by treatment arm
totals <- adsl_f         \%>\%
  mutate(TRTP = "Total") \%>\%
  bind_rows(adsl_f)      \%>\%
  group_by(TRTP)         \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate the summary statistcs
output <- adlb                                                  \%>\%
  mutate(TRTP = "Total")                                        \%>\%
  bind_rows(adlb)                                               \%>\%
  filter(SAFFL == "Y" & PARAMCD \%in\% c("ABSR148N", "ABTIT148")) \%>\%
  group_by(TRTP, PARAM, STUDYID)                                \%>\%
  summarize(N  = n_distinct(USUBJID),
            Mean = mean(AVAL, na.rm = TRUE),
            SD   = sd(  AVAL, na.rm = TRUE),
            Min  = min( AVAL, na.rm = TRUE),
            Max  = max( AVAL, na.rm = TRUE)
  )

# The NNTable is generated in small steps to view what happens

# Initiate the table
.NNTable <- NNTable(output, "PARAM", "TRTP", "N", "Mean (SD)", "Min ; Max")

# View the table
.NNTable
.NNTable \%>\% addPrinter(type = "flextable")


# Transpose the summary columns to long format
.NNTable <- .NNTable                                    \%>\%
  addTransLong(SUM = c("N", "Mean (SD)" , "Min ; Max"),
               var_name = "Called var")                 \%>\%
  addFormat(format_data = c(N = "\%.0f"), dec = 2)

# View the table
.NNTable

# Add exposure, transpose to wide and ad grouping variables
.NNTable <- addExposure(.NNTable, exposure = totals,
                        format_alone = TRUE,
                        format_data  = c("Number of subjects" = "\%.0f")) \%>\%
            addTransWide(TRTP = "SUM")                                   \%>\%
            addGroupedColumns("PARAM", "Called var")

# View the final result
.NNTable

.NNTable \%>\% addPrinter(type = "flextable")


## *****************************************************************************
## # Example of transpose from wide to long                                 ----
##
## Calculate the statistics for an AE table
## *****************************************************************************

library(dplyr)
#library(NNR)

# Filter data according Safety analysis set
adsl_f <- adsl            \%>\%
  filter(SAFFL == "Y")    \%>\%
    rename(TRTA = TRT01P)

# Calculate total exposure
totals <- adsl_f        \%>\%
  group_by(TRTA)        \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate adverse event statistics
output <- adae \%>\% group_by(AEBODSYS, AEDECOD, TRTA) \%>\%
  dplyr::summarise(N = n_distinct(USUBJID),
                   P = N / nrow(adsl),
                   E = n())

# Generate the NNTable ------------------------------------------------------

# Initiate by defining the columns based on ADAE
.NNTable <- NNTable(output, "AEBODSYS", "AEDECOD", "TRTA", "N", "(\%)" = "(P)", "E")

# Add exposure calculated from ADSL
.NNTable <- .NNTable \%>\%
  addExposure(exposure = totals,
              format_alone = FALSE,
              format_data  = c("Number of subjects"     = "\%.0f",
                               "Total exposure (years)" = "\%.2f"))

# Transpose to wide
.NNTable <- .NNTable \%>\% addTransWide(TRTA = c("N", "(\%)", "E"))

# Add Filling such that non-observed events is assigned 0
.NNTable <- .NNTable \%>\% addFilling(N = 0)

# Add formatting information for the table
.NNTable <- .NNTable \%>\% addFormat(format_data = c(N = "\%.0f", P = "\%.2f", E = "\%.0f"))

# Group the columns AEBODSYS and AEDECOD
.NNTable <- .NNTable \%>\% addGroupedColumns("AEBODSYS", "AEDECOD" , name = "some name")

# Truncate columns at a certain width such that long names appear on multiple lines
.NNTable <- .NNTable \%>\% addTruncation(AEBODSYS = 50, AEDECOD = 48)

# View the final result
.NNTable

.NNTable \%>\% addPrinter(type = "flextable")

}
