% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSpaceControl.R
\name{addSpaceControl}
\alias{addSpaceControl}
\title{Add specifications to the width beteen columns}
\usage{
addSpaceControl(
  .NNTable,
  spread = TRUE,
  max.sep = 50,
  max.space = 50,
  sep = 1,
  space = 1
)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}
should the leading space be discarded}

\item{spread}{\code{logical} Should the columns be spread out to fill the page width}

\item{max.sep}{\code{integer} indicating maximum with between regular columns}

\item{max.space}{\code{integer} indicating maximum with between grouped columns}

\item{sep}{\code{integer} indicating the width of the separators between columns}

\item{space}{\code{integer} indicating the width of the separation between grouped columns}
}
\value{
An object of class \code{NNTable} with the page split added
}
\description{
Add specifications to the width beteen columns
}
