% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTranslate.R
\name{addTranslate}
\alias{addTranslate}
\title{Add translations to an NNTable}
\usage{
addTranslate(.NNTable, ...)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{...}{the columns in data that needs to be translated. e.g
\code{SEX = c("Male" = "M", "Female" = "F")} will replace \code{"M"} by \code{"Male"}.
Alternatively, it is possible to pass a function, e.g. to transform column
\code{AESOC} to title case  \code{"AESOC" = stringr::str_to_title}.
The column names should be the original column names and should not
reflect any renaming.}
}
\value{
An object of class \code{NNTable} with the columns translated
}
\description{
Sometimes the values in data are not very informative and needs to be translated,
e.g. for SEX: "Female" instead of "F" and "Male" instead of "M".
}
\examples{
library(NNR)
library(NNBiostat)

# make sure the training data is up to date
NNtraining::createTrainingDB()
(db <- nnaccess("0001", root = "~/training"))

adae_f <- db$adam("adae") \%>\% filter(SAFFL == "Y" & TRTA != "")
adsl_f <- db$adam("adsl") \%>\% filter(SAFFL == "Y" & TRT30A != "") \%>\%
  rename(TRTA = TRT30A)


subjects <- adsl_f \%>\%
  group_by(TRTA, SEX) \%>\%
  summarise("Number of subjects" = n_distinct(USUBJID),
            "Total exposure" = sum(TRDURY),
            .groups = "drop")

subjects$SEX <- as.factor(subjects$SEX)


stats <- adae_f \%>\%
  group_by(TRTA,AESOC,SEX) \%>\%
  summarise(N = n_distinct(USUBJID),
            E = n(), .groups = "drop")

stats$SEX <- as.factor(stats$SEX)
table <- NNTable(stats, "TRTA", Sex = "SEX", "AESOC","N","E") \%>\%
  addTransWide("TRTA" = c("N","E")) \%>\%
  addGroupedColumns("AESOC") \%>\%
  addOrder(E = -1)

# print the table
table

# we would like to rename some values to make the content more clear
table \%>\%
  addTranslate(TRTA  = c("Ico" = "A", "NOVO rapid" = "B"),
               SEX   = c("Female" = "F", "Male" = "M"),
               AESOC = stringr::str_to_title)

}
