% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTransLong.R
\name{addTransLong}
\alias{addTransLong}
\title{Add columns that needs to be transposed from wide to long}
\usage{
addTransLong(
  .NNTable,
  ...,
  value_name,
  var_name = "name",
  display_name,
  remove_blank = TRUE
)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{...}{\code{character} or \code{list} specifying the columns to
transpose.The names of the character will be the names displayed in the
table when transposed to long.}

\item{value_name}{\code{character} with the name of the value column}

\item{var_name}{\code{character} with the name of the name column}

\item{display_name}{The display name of the name column}

\item{remove_blank}{\code{logical} Should blank rows be deleted. If only some
should be deleted add a space to the specific row and column combination}
}
\value{
An object of class \code{NNTable} with the transpose specified
}
\description{
Add columns that needs to be transposed from wide to long
}
\examples{

## *****************************************************************************
## # Calculate summary statistics                                           ----
##
## The exposure/summary statistics are calculated for ADSL and ADLB
## *****************************************************************************

# filter and rename adsl to get totals
# library(NNR)
library(dplyr)

adsl_f <- adsl       \%>\%
filter(FASFL == "Y") \%>\%
  rename(TRTP = TRT01P)

# calculate the totals by treatment arm
totals <- adsl_f         \%>\%
  mutate(TRTP = "Total") \%>\%
  bind_rows(adsl_f)      \%>\%
  filter(FASFL == "Y")   \%>\%
  group_by(TRTP)         \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate the summary statistcs
output <- adlb                                                  \%>\%
  mutate(TRTP = "Total")                                        \%>\%
  bind_rows(adlb)                                               \%>\%
  filter(SAFFL == "Y" & PARAMCD \%in\% c("ABSR148N", "ABTIT148")) \%>\%
  group_by(TRTP, PARAM, STUDYID)                                \%>\%
  summarize(N  = n_distinct(USUBJID),
            Mean = mean(AVAL, na.rm = TRUE),
            SD   = sd(  AVAL, na.rm = TRUE),
            Min  = min( AVAL, na.rm = TRUE),
            Max  = max( AVAL, na.rm = TRUE)
  )

# The NNTable is generated in small steps to view what happens

# Initiate the table
.NNTable <- NNTable(output, "PARAM", "TRTP", "N", "Mean (SD)", "Min ; Max")

# View the table
.NNTable


# Transpose the summary columns to long format
.NNTable <- .NNTable                                    \%>\%
  addTransLong(SUM = c("N", "Mean (SD)" , "Min ; Max"),
               var_name = "Called var")                 \%>\%
  addFormat(format_data = c(N = "\%.0f"), dec = 2)

# View the table
.NNTable

# Add exposure, transpose to wide and ad grouping variables
.NNTable <- addExposure(.NNTable, exposure = totals,
                        format_alone = TRUE,
                        format_data  = c("Number of subjects" = "\%.0f")) \%>\%
            addTransWide(TRTP = "SUM")                                   \%>\%
            addGroupedColumns("PARAM", "Called var")

# View the final result
.NNTable

}
