% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUnderScore.R
\name{addUnderScore}
\alias{addUnderScore}
\title{Add underscore to an NNTable header}
\usage{
addUnderScore(.NNTable, add = TRUE)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by
\code{\link{NNTable}}}

\item{add}{\code{logical} indicating whether the underscore should be added}
}
\value{
An object of class \code{NNTable} where underscores will be made in
the header
}
\description{
Add underscore to an NNTable header
}
