% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_add.R
\name{addWrapping}
\alias{addWrapping}
\title{Add wrapping to an NNTable}
\usage{
addWrapping(
  .NNTable,
  title,
  footnote,
  sys_footnote,
  title_space = "",
  footer_space = ""
)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{title}{\code{character} with the title}

\item{footnote}{\code{character} with the footnote}

\item{sys_footnote}{\code{character} with the system footnote}

\item{title_space}{\code{character} to be added between the title and
and table header}

\item{footer_space}{\code{character} to be added between the footnote and
sys_footnote.}
}
\value{
An object of class \code{NNTable} with the wrapping specified
}
\description{
Add wrapping to an NNTable
}
\examples{

## *****************************************************************************
## # Example of transpose from wide to long                                 ----
##
## Calculate the statistics for an AE table
## *****************************************************************************

library(dplyr)
# library(NNR)

# Filter data according Safety analysis set
adsl_f <- adsl            \%>\%
  filter(SAFFL == "Y")    \%>\%
    rename(TRTA = TRT01P)

# Calculate total exposure
totals <- adsl_f        \%>\%
  group_by(TRTA)        \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate adverse event statistics
output <- adae \%>\% group_by(AEBODSYS, AEDECOD, TRTA) \%>\%
  dplyr::summarise(N = n_distinct(USUBJID),
                   P = N / nrow(adsl),
                   E = n())

# Generate the NNTable ------------------------------------------------------

# Initiate by defining the columns based on ADAE
.NNTable <- NNTable(output, "AEBODSYS", "AEDECOD", "TRTA", "N", "(\%)" = "(P)", "E") \%>\%
            addTruncation(AEBODSYS = 50, AEDECOD = 48)                              \%>\%
            addGroupedColumns("AEBODSYS", "AEDECOD" , name = "some name")

# Add the title and footnote, these can span several lines and will
#  automatically be wrapped
.NNTable <- .NNTable \%>\%
            addWrapping(title        = "This could be a very long title",
                        footnote     = "This could be a very long footnote",
                        sys_footnote = "Explain where the data comes from")

# View the final result
.NNTable
}
