% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlignment.R
\name{align}
\alias{align}
\title{Align text in monospaced output}
\usage{
align(
  x,
  alignment = c("left", "center", "right"),
  width = max(nchar(text)),
  type = c("non-trimmed", "trimmed"),
  keep.empty = TRUE
)
}
\arguments{
\item{x}{\code{character} to center}

\item{alignment}{the type of alignment.}

\item{width}{the number of characters of the column}

\item{type}{should text be trimmed or not}

\item{keep.empty}{should completely empty comlumns be left blank}
}
\value{
The \code{character} \code{x} with added spaces
}
\description{
Align text in monospaced output
}
