% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlignment.R
\name{alignCenter}
\alias{alignCenter}
\title{Center text in monospaced output}
\usage{
alignCenter(
  text,
  width = max(nchar(text)),
  type = c("non-trimmed", "trimmed"),
  fill = TRUE,
  keep.empty = TRUE
)
}
\arguments{
\item{text}{\code{character} to center}

\item{width}{the number of characters of the column}

\item{type}{should text be trimmed or not}

\item{fill}{Should trailing blanks be added}

\item{keep.empty}{should completely empty comlumns be left blank}
}
\value{
The \code{character} \code{text} with added spaced in front
}
\description{
Center text in monospaced output
}
\keyword{internal}
