% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlignment.R
\name{alignLeft}
\alias{alignLeft}
\title{Add space on right for monospaced output}
\usage{
alignLeft(
  text,
  width = max(nchar(text)),
  type = c("non-trimmed", "trimmed"),
  sep = " ",
  keep.empty = TRUE
)
}
\arguments{
\item{text}{\code{character} to center}

\item{width}{the number of characters of the column}

\item{type}{should text be trimmed or not}

\item{sep}{The inserted space generator, defaults to " ".}

\item{keep.empty}{should completely empty comlumns be left blank}
}
\value{
The \code{character} \code{text} wtihout leading and trailing blanks
}
\description{
Add space on right for monospaced output
}
\keyword{internal}
