% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_misc.R
\name{alignRight}
\alias{alignRight}
\title{Add space on left for monospaced output}
\usage{
alignRight(
  text,
  width = max(nchar(text)),
  type = c("non-trimmed", "trimmed"),
  keep.empty = TRUE
)
}
\arguments{
\item{text}{\code{character} to center}

\item{width}{the number of characters of the column}

\item{type}{should text be trimmed or not}

\item{keep.empty}{should completely empty comlumns be left blank}
}
\value{
The \code{character} \code{text} wtihout leading and trailing blanks
}
\description{
Add space on left for monospaced output
}
