% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_apply.R
\name{apply_createTree}
\alias{apply_createTree}
\title{Create a tree structure of the table}
\usage{
apply_createTree(.NNTable)
}
\arguments{
\item{.NNTable}{An \code{NNTable} object}
}
\value{
The \code{NNTable} object with the tree added
}
\description{
Create a tree structure of the table
}
