% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkEmpty.R
\name{checkEmpty}
\alias{checkEmpty}
\title{Replace an NNTable object with an empty one if no data is present}
\usage{
checkEmpty(.NNTable)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}
}
\value{
An object of class \code{NNTable} without any data
}
\description{
Replace an NNTable object with an empty one if no data is present
}
\keyword{internal}
