% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_misc.R
\name{depth}
\alias{depth}
\title{Get the depth of a list}
\usage{
depth(list, current_depth = 0)
}
\arguments{
\item{list}{The list to determine the with}

\item{current_depth}{The starting value of depth}
}
\value{
\code{numeric} indicating the deepest neting level of the list
}
\description{
Get the depth of a list
}
