% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandTotals.R
\name{expandTotals}
\alias{expandTotals}
\title{expand with Totals for specific columns in a tibble}
\usage{
expandTotals(.data, var, name = rep("Total", length(var)))
}
\arguments{
\item{.data}{data that needs to have totals added}

\item{var}{column names for which a total needs to be added}

\item{name}{Name the value that is assigned as the total}
}
\value{
a tibble with the new rows added
}
\description{
Add the totals for specified groups of variables (i.e. "var")
Note, the cells will be populated with the text in "name"
}
\examples{
# library(dplyr)
# summary_data <- adsl \%>\% filter(SAFFL == "Y") \%>\% expandTotals(c("TRT01A", "COUNTRY2")) \%>\% 
#     group_by_at(c("TRT01A", "COUNTRY2")) \%>\% summarize(N = n_distinct(USUBJID))
}
