% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_misc.R
\name{Format.default}
\alias{Format.default}
\title{Format default}
\usage{
\method{Format}{default}(x, ...)
}
\arguments{
\item{x}{The vector to format}

\item{...}{additional arguments passed on}
}
\value{
The formatted vector
}
\description{
Format default
}
