% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_misc.R
\name{Format}
\alias{Format}
\title{Format numeric}
\usage{
Format(x, ...)
}
\arguments{
\item{x}{The vector to format}

\item{...}{additional arguments passed on}
}
\value{
The formatted vector
}
\description{
Format numeric
}
\examples{
Format(c("ad", "chrjf"))
Format(c(rnorm(10, 10, 20)))
}
