% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_misc.R
\name{hline}
\alias{hline}
\title{Line to separate header and footnote from table}
\usage{
hline(symbol = "horizontal", times = 141)
}
\arguments{
\item{symbol}{\code{character} symbol for the line}

\item{times}{the number of times the \code{symbol} should be repeated.}
}
\value{
the horizontal line
}
\description{
Line to separate header and footnote from table
}
\keyword{internal}
