% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable.R
\name{NNTable}
\alias{NNTable}
\title{Initialise an NN table}
\usage{
NNTable(.data, ..., .font_size = 8, .orientation = "port", page_size = NULL)
}
\arguments{
\item{.data}{\code{data.frame} to output as table}

\item{...}{Names of the columns in .data that should be included in the
table. If none are suppied all columns are used. If named, e.g. "New name"
= "old name" the new name will be displayed.}

\item{.font_size}{\code{numeric} The font size to use. For the \code{"txt"} print it must be 8, 9, or 10.}

\item{.orientation}{\code{character} The orientation of the page, \code{"port"} for portrait and \code{"land"} for landscape.}

\item{page_size}{\code{list} containing page.length and page.width dictating
the number of lines and number of characters on the page}
}
\value{
An object of type \code{NNTable}
}
\description{
Initialise an NN table
}
\details{
In order to change the presentation of data several "add" functions
have been made
\itemize{
\item \link{addAlignment}
\itemize{
\item Add alignment to the columns, can be center = "c", left = "l", and right = "r"
}
\item \link{addCellSplit}
\itemize{
\item Add cell split in order to divide one cell of a table into multiple rows, i.e. add a newline
}
\item \link{addExposure}
\itemize{
\item Add exposure information such as total number of subjects and years of exposure.
}
\item \link{addFilling}
\itemize{
\item When transposing from long to wide some of the columns may become empty.
This function is used to fill those empty columns with a value, eg setting
N = 0 for adverse events where no events were observed in a specific arm.
}
\item \link{addFormat}
\itemize{
\item Add formatting info to an NNTable, e.g. number of decimals or specific
formatting for each cell of the table
}
\item \link{addGroupedColumns}
\itemize{
\item Add grouped columns to an NNTable
}
\item \link{addOrder}
\itemize{
\item Add ordering variables for NNTable
}
\item \link{addTransLong}
\itemize{
\item Add columns that needs to be transposed from wide to long
}
\item \link{addTransWide}
\itemize{
\item Add columns that needs to be transposed from long to wide
}
\item \link{addTruncation}
\itemize{
\item Truncate columns at specified width. If the columns is longer than the
specified width it will continue on the following line with two spaces added
}
\item \link{addUnderScore}
\itemize{
\item Add underscores in the header of NNTable
}
\item \link{addWrapping}
\itemize{
\item Add title and footnote to the table. These are usually added using the
\code{exportOutput} function stored within the \code{nnaccess} object created by the function \code{nnaccess}
}
}
}
\examples{
## *****************************************************************************
## # Calculate summary statistics                                           ----
##
## The exposure/summary statistics are calculated for ADSL and ADLB
## *****************************************************************************

# filter and rename adsl to get totals
# library(NNR)
library(dplyr)

adsl_f <- adsl       \%>\%
filter(FASFL == "Y") \%>\%
  rename(TRTP = TRT01P)

# calculate the totals by treatment arm
totals <- adsl_f         \%>\%
  mutate(TRTP = "Total") \%>\%
  bind_rows(adsl_f)      \%>\%
  group_by(TRTP)         \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate the summary statistcs
output <- adlb                                                  \%>\%
  mutate(TRTP = "Total")                                        \%>\%
  bind_rows(adlb)                                               \%>\%
  filter(SAFFL == "Y" & PARAMCD \%in\% c("ABSR148N", "ABTIT148")) \%>\%
  group_by(TRTP, PARAM, STUDYID)                                \%>\%
  summarize(N  = n_distinct(USUBJID),
            Mean = mean(AVAL, na.rm = TRUE),
            SD   = sd(  AVAL, na.rm = TRUE),
            Min  = min( AVAL, na.rm = TRUE),
            Max  = max( AVAL, na.rm = TRUE)
  )

# The NNTable is generated in small steps to view what happens

# Initiate the table
.NNTable <- NNTable(output, "PARAM", "TRTP", "N", "Mean (SD)", "Min ; Max")

# View the table
.NNTable


# Transpose the summary columns to long format
.NNTable <- .NNTable                                    \%>\%
  addTransLong(SUM = c("N", "Mean (SD)" , "Min ; Max"),
               var_name = "Called var")                 \%>\%
  addFormat(format_data = c(N = "\%.0f"), dec = 2)

# View the table
.NNTable

# Add exposure, transpose to wide and ad grouping variables
.NNTable <- addExposure(.NNTable, exposure = totals,
                        format_alone = TRUE,
                        format_data  = c("Number of subjects" = "\%.0f")) \%>\%
            addTransWide(TRTP = "SUM")                                   \%>\%
            addGroupedColumns("PARAM", "Called var")

# View the final result
.NNTable



## *****************************************************************************
## # Example of transpose from wide to long                                 ----
##
## Calculate the statistics for an AE table
## *****************************************************************************

library(dplyr)
#library(NNR)

# Filter data according Safety analysis set
adsl_f <- adsl            \%>\%
  filter(SAFFL == "Y")    \%>\%
    rename(TRTA = TRT01P)

# Calculate total exposure
totals <- adsl_f        \%>\%
  group_by(TRTA)        \%>\%
  summarize("Number of subjects"     = n_distinct(USUBJID),
            "Total exposure (years)" = sum(TR01DURY))

# Calculate adverse event statistics
output <- adae \%>\% group_by(AEBODSYS, AEDECOD, TRTA) \%>\%
  dplyr::summarise(N = n_distinct(USUBJID),
                   P = N / nrow(adsl),
                   E = n())

# Generate the NNTable ------------------------------------------------------

# Initiate by defining the columns based on ADAE
.NNTable <- NNTable(output, "AEBODSYS", "AEDECOD", "TRTA", "N", "(\%)" = "(P)", "E")

# Add exposure calculated from ADSL
.NNTable <- .NNTable \%>\%
  addExposure(exposure = totals,
              format_alone = FALSE,
              format_data  = c("Number of subjects"     = "\%.0f",
                               "Total exposure (years)" = "\%.2f"))

# Transpose to wide
.NNTable <- .NNTable \%>\% addTransWide(TRTA = c("N", "(\%)", "E"))

# Add Filling such that non-observed events is assigned 0
.NNTable <- .NNTable \%>\% addFilling(N = 0)

# Add formatting information for the table
.NNTable <- .NNTable \%>\% addFormat(format_data = c(N = "\%.0f", P = "\%.2f", E = "\%.0f"))

# Group the columns AEBODSYS and AEDECOD
.NNTable <- .NNTable \%>\% addGroupedColumns("AEBODSYS", "AEDECOD" , name = "some name")

# Truncate columns at a certain width such that long names appear on multiple lines
.NNTable <- .NNTable \%>\% addTruncation(AEBODSYS = 50, AEDECOD = 48)

# View the final result
.NNTable
}
