% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable.R
\name{print.NNTable}
\alias{print.NNTable}
\title{Print an NNTable object}
\usage{
\method{print}{NNTable}(x, ..., page = 1, file = NULL, verbose = TRUE, check_empty = TRUE)
}
\arguments{
\item{x}{The NNTable object}

\item{...}{Additional arguments currently not used}

\item{page}{The page to display}

\item{file}{\code{character} optional for writing output to a file}

\item{verbose}{when the print is used to generate formatting results it is not
desired to print the result}

\item{check_empty}{if TRUE data sets with zero rows are output as empty}
}
\value{
prints the table
}
\description{
Print an NNTable object
}
