% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_read_encoded.R
\name{read_encoded}
\alias{read_encoded}
\title{Read an encoded txt output}
\usage{
read_encoded(file, enc = "CP1252")
}
\arguments{
\item{file}{\code{character} indicating the file to read}

\item{enc}{\code{character} indicating the encoding to be used. The EOT tool
expects this to be in CP1252}
}
\value{
character with the read lines
}
\description{
Read an encoded txt output
}
