% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_sub.R
\name{subsetting}
\alias{subsetting}
\alias{[.NNTable}
\title{Subsetting NNTables}
\usage{
\method{[}{NNTable}(x, i, j, data = c("main", "exposure"))
}
\arguments{
\item{x}{An NNtable}

\item{i}{Row subset}

\item{j}{Column subset}

\item{data}{\code{character} the data to subset, Can be either \code{"main"},
\code{"exposure"} or both  \code{c("main", "exposure")} to subset the associated data set}
}
\value{
An NNTable subset as specified
}
\description{
Subset a created NNTable so that only some columns are shown.
}
\examples{
library(NNR)
library(NNBiostat)

# make sure the training data is up to date
NNtraining::createTrainingDB()
(db <- nnaccess("0001", root = "~/training"))

adae_f <- db$adam("adae") \%>\% filter(SAFFL == "Y" & TRTA != "")
adsl_f <- db$adam("adsl") \%>\% filter(SAFFL == "Y" & TRT30A != "") \%>\%
 rename(TRTA = TRT30A)


subjects <- adsl_f \%>\%
 group_by(TRTA, SEX) \%>\%
 summarise("Number of subjects" = n_distinct(USUBJID),
           "Total exposure" = sum(TRDURY),
           .groups = "drop")

stats <- adae_f \%>\%
  group_by(TRTA,AESOC,SEX) \%>\%
  summarise(AGE = mean(AGE),
            N = n_distinct(USUBJID),
            E = n(), .groups = "drop")

table <- NNTable(stats, "TRTA", sex = "SEX", "AGE", "AESOC","N","E") \%>\%
  addTransWide("TRTA" = c("N","E")) \%>\%
  addGroupedColumns("AESOC") \%>\%
  addOrder(E = -1) \%>\%
  addTranslate(TRTA  = c("Ico" = "A", "NOVO rapid" = "B"),
               SEX   = c("Female" = "F", "Male" = "M"),
               AESOC = stringr::str_to_title)


# print the table
table

# Subset to only females
table[SEX == "F"]

# Remove the column age
table[, -"AGE"]

# Adding an exposure table
exp_table <- table \%>\% addExposure(subjects)

exp_table

# subset to only females
exp_table[SEX == "F"]

# subset and remove the columns SEX and AGE
# here it is important to note that AGE is only present in the main data

exp_table[SEX == "F", -"SEX"][, -"AGE", "main"]

# since AGE is not contained in the exposure data we need to instruct the
# subset to only look in the main data

}
