% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_read_encoded.R
\name{write_encoded}
\alias{write_encoded}
\title{Write a file readeble by EOT}
\usage{
write_encoded(text, file, enc = "CP1252")
}
\arguments{
\item{text}{\code{character} where each element represent a line in the final
output}

\item{file}{\code{character} indicating the full name of the output}

\item{enc}{\code{character} indicating the encoding to be used. The EOT tool
expects this to be in CP1252}
}
\value{
The encoded string is returned invisible
}
\description{
Write a file readeble by EOT
}
