% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPageSplit.R
\name{addPageSplit}
\alias{addPageSplit}
\title{Add grouped columns to an NNTable to form the table STUB}
\usage{
addPageSplit(
  .NNTable,
  keepers = character(0),
  cuts = character(0),
  drop_intial_space = TRUE
)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}

\item{keepers}{\code{character} with the names of the columns to keep on all pages.}

\item{cuts}{columns that should be moved to the following page}

\item{drop_intial_space}{\code{logical} When splitting a table with no keepers,
should the leading space be discarded}
}
\value{
An object of class \code{NNTable} with the page split added
}
\description{
Add grouped columns to an NNTable to form the table STUB
}
