% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFilling.R
\name{apply_filling}
\alias{apply_filling}
\title{Add values for colums that are not present in data}
\usage{
apply_filling(.NNTable)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}
}
\description{
Add values for colums that are not present in data
}
\keyword{internal}
