% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_apply.R
\name{apply_format_concat}
\alias{apply_format_concat}
\title{Apply specified format and concatenate columns}
\usage{
apply_format_concat(.NNTable)
}
\arguments{
\item{.NNTable}{An \code{NNTable} generated by \code{\link{NNTable}}}
}
\description{
Apply specified format and concatenate columns
}
