% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPageSplit.R
\name{balanced.split}
\alias{balanced.split}
\title{Main function that splits all.values into n.splits
chunks, minimizing sd(sum(chunk))}
\usage{
balanced.split(values, n.splits, p_width = Inf)
}
\description{
Main function that splits all.values into n.splits
chunks, minimizing sd(sum(chunk))
}
