% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNTable_misc.R
\name{Format.integer}
\alias{Format.integer}
\title{Format numeric}
\usage{
\method{Format}{integer}(x, ..., format = guessFormat(x, dec = min(3, decimalPlaces(x))))
}
\arguments{
\item{x}{\code{numeric} to}

\item{...}{additional arguments passed on}

\item{format}{The format to use}
}
\value{
The formated vector
}
\description{
Format numeric
}
